requirements.txt
pony
pillow
Mastodon.py